+++ include/aws/s3/private/s3_endpoint_resolver.h
@@ -1,0 +1,12 @@
+#ifndef AWS_S3_ENDPOINT_RESOLVER_PRIVATE_H
+#define AWS_S3_ENDPOINT_RESOLVER_PRIVATE_H
+
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
+extern const char aws_s3_endpoint_resolver_partitions[];
+extern const char aws_s3_endpoint_rule_set[];
+
+#endif /* AWS_S3_ENDPOINT_RESOLVER_PRIVATE_H */
