+++ source/s3_endpoint_resolver/s3_endpoint_resolver.c
@@ -3,30 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
-#include <aws/common/ref_count.h>
-#include <aws/s3/s3_endpoint_resolver/s3_endpoint_resolver.h>
+#include <aws/s3/s3_endpoint_resolver.h>
+#include <aws/sdkutils/endpoints_rule_engine.h>
 #include <aws/sdkutils/partitions.h>
 
-struct aws_s3_endpoint_resolver {
-    struct aws_allocator *allocator;
-    struct aws_ref_count ref_count;
-    struct aws_endpoints_rule_engine *rule_engine;
-};
-
-static void s_aws_s3_endpoint_resolver_clean_up(struct aws_s3_endpoint_resolver *endpoint_resolver) {
-    aws_endpoints_rule_engine_release(endpoint_resolver->rule_engine);
-    aws_mem_release(endpoint_resolver->allocator, endpoint_resolver);
-}
-
-struct aws_s3_endpoint_resolver *aws_s3_endpoint_resolver_new(struct aws_allocator *allocator) {
+struct aws_endpoints_rule_engine *aws_s3_endpoint_resolver_new(struct aws_allocator *allocator) {
     struct aws_endpoints_ruleset *ruleset = NULL;
     struct aws_partitions_config *partitions = NULL;
-    struct aws_s3_endpoint_resolver *resolver = aws_mem_calloc(allocator, 1, sizeof(struct aws_s3_endpoint_resolver));
-    bool success = false;
-
-    resolver->allocator = allocator;
-    aws_ref_count_init(
-        &resolver->ref_count, resolver, (aws_simple_completion_callback *)s_aws_s3_endpoint_resolver_clean_up);
+    struct aws_endpoints_rule_engine *rule_engine = NULL;
 
     struct aws_byte_cursor ruleset_cursor = aws_byte_cursor_from_c_str(aws_s3_endpoint_rule_set);
     struct aws_byte_cursor partitions_cursor = aws_byte_cursor_from_c_str(aws_s3_endpoint_resolver_partitions);
@@ -41,43 +25,20 @@ struct aws_s3_endpoint_resolver *aws_s3_endpoint_resolver_new(struct aws_allocat
         goto cleanup;
     }
 
-    resolver->rule_engine = aws_endpoints_rule_engine_new(allocator, ruleset, partitions);
-    if (!resolver->rule_engine) {
-        goto cleanup;
-    }
-
-    success = true;
+    rule_engine = aws_endpoints_rule_engine_new(allocator, ruleset, partitions);
 
 cleanup:
     aws_endpoints_ruleset_release(ruleset);
     aws_partitions_config_release(partitions);
-    if (!success) {
-        s_aws_s3_endpoint_resolver_clean_up(resolver);
-        resolver = NULL;
-    }
-    return resolver;
-}
-
-struct aws_s3_endpoint_resolver *aws_s3_endpoint_resolver_acquire(struct aws_s3_endpoint_resolver *endpoint_resolver) {
-    if (endpoint_resolver != NULL) {
-        aws_ref_count_acquire(&endpoint_resolver->ref_count);
-    }
-    return endpoint_resolver;
-}
-
-struct aws_s3_endpoint_resolver *aws_s3_endpoint_resolver_release(struct aws_s3_endpoint_resolver *endpoint_resolver) {
-    if (endpoint_resolver != NULL) {
-        aws_ref_count_release(&endpoint_resolver->ref_count);
-    }
-    return endpoint_resolver;
+    return rule_engine;
 }
 
 struct aws_endpoints_resolved_endpoint *aws_s3_endpoint_resolver_resolve_endpoint(
-    struct aws_s3_endpoint_resolver *resolver,
+    struct aws_endpoints_rule_engine *rule_engine,
     struct aws_endpoints_request_context *request_context) {
 
     struct aws_endpoints_resolved_endpoint *resolved_endpoint = NULL;
-    aws_endpoints_rule_engine_resolve(resolver->rule_engine, request_context, &resolved_endpoint);
+    aws_endpoints_rule_engine_resolve(rule_engine, request_context, &resolved_endpoint);
 
     return resolved_endpoint;
 }
